+++ scripts/build.ps1
@@ -91,10 +91,12 @@ copy_app "org.nebulaservices.texteditor"
 copy_app "org.nebulaservices.repostore"
 copy_app "org.nebulaservices.files"
 copy_app "org.nebulaservices.browser"
+copy_app "org.nebulaservices.processmanager"
+copy_app "xen.runtime"  
 
 Copy-Item -Recurse ./apps/ ./build/apps/
 
-success "Copied apps"
+success "Copied apps and libs"
 section "Building Workbox libraries"
 
 New-Item -ItemType Directory -Force build/libs/workbox/ | Out-Null